requirements.txt
Flask==3.0.0
pytest==7.4.0
python-dotenv==1.0.1